backend/requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-multipart
pyjwt
